Ensure sensors are created for RF, DTD, and Wireguard links
We dynamically create sensors to count links of each type, but let's
guarantee we always create them for the common link types, so at least
we provide a sensor with a value of '0' if we just didn't see any of that
link when we polled the node. We'll start with:
RF
DTD
WIREGUARD

It's one line to add another later.